+++ CHANGELOG.md
@@ -50,5 +50,6 @@
 - `BuildpackApi` no longer implements `FromStr`, use `BuildpackApi::try_from()` instead.
 - Fixed file extension for delimiters when writing `LayerEnv` to disk.
 - Fixed the `group` field on `buildpack::Order` to now be public.
+- Add an external Cargo command for packaging libcnb buildpacks. See the README for usage.
 
 ## [0.3.0] 2021/09/17